pyproject.toml
[project]
name = "iot-trabalho-final"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy==2.3.4",
    "pandas==2.3.3",
    "pydantic==2.12.3",
    "requests==2.32.5",
    "uvicorn==0.38.0",
    "ultralytics==8.0.227",
    "grpcio==1.76.0",
    "grpcio-tools==1.76.0",
    "protobuf==6.33.0",
    "torch==2.9.0",
]
